Fixed A Bit of memory Leaks

+++ ChronoLoop/ChronoLoop/Actions/CCLevel1Tutorial.h
@@ -224,7 +224,7 @@ namespace Epoch {
 					((SFXEmitter*)boards[2]->GetComponentIndexed(eCOMPONENT_AUDIOEMITTER, 0))->CallEvent();
 
 					if (TimeManager::Instance()->GetCurrentSnapFrame() < 60)
-						Settings::GetInstance().SetFloat("TutorialRewind - FinalProgress", TimeManager::Instance()->GetCurrentSnapFrame());
+						Settings::GetInstance().SetFloat("TutorialRewind - FinalProgress", (float)TimeManager::Instance()->GetCurrentSnapFrame());
 					else
 						Settings::GetInstance().SetFloat("TutorialRewind - FinalProgress", 60);
 					scalingDone = false;
